requirements.txt
pandas>=2.0
numpy>=2.0
openpyxl>=3.1
duckdb>=1.0
langgraph>=1.1
langchain-core>=1.2
langchain-google-genai>=4.2
python-dotenv>=1.0
pydantic>=2.10
fastapi>=0.135
uvicorn>=0.43
pyyaml>=6.0
matplotlib>=3.10
seaborn>=0.13
python-multipart
